Images of each Pokemon now displays. Generation 1 - 5 can now be displayed.
The next thing I want to work on is providing more information for each pokemon.
