+++ src/routes/about.svelte
@@ -1,1 +1,3 @@
 <h1 class="text-4xl text-center my-8 uppercase">About Svelte & Pokedex</h1>
+<p>The Pokedex gets the data from <a class="underline underline-offset-1" target="_target" href="https://pokeapi.co/">PokeAPI</a>.</p>
+<p>The platform is built on SvelteKit.</p>
